updated tests

+++ api/controllers/lists.js
@@ -50,10 +50,13 @@ const deleteList = wrapAsync(async (request, response) => {
   // Remove the list and all of the tasks on it
   // TODO: Convert to schema middleware (listSchema.pre('remove', ...))
   const deletedList = await List.findByIdAndRemove(request.params.id);
-  await Task.deleteMany({ list: deletedList._id });
 
-  // Remove the list's id from the User.lists array it was in
-  await User.findByIdAndUpdate(deletedList.user, { $pull: { lists: deletedList._id } });
+  if (deletedList) {
+    await Task.deleteMany({ list: deletedList._id });
+
+    // Remove the list's id from the User.lists array it was in
+    await User.findByIdAndUpdate(deletedList.user, { $pull: { lists: deletedList._id } });
+  }
 
   // Return '204 No Content' in all cases
   response.status(204).end();
